requirement.txt
Flask
Flask-Cors
PyPDF2
pyttsx3
python-dotenv
google-generativeai
gunicorn # For production deployment